+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v2.6.0
+## 05/22/2017
+
+1. [](#improved)
+    * Inherit options from plugin configuration [#39](https://github.com/getgrav/grav-plugin-email/pull/39)
+1. [](#bugfix)
+    * Also process translation on the email subject [https://github.com/getgrav/grav-plugin-comments/issues/38](https://github.com/getgrav/grav-plugin-comments/issues/38)
+
 # v2.5.3
 ## 01/03/2017